LayoutTests/platform/gtk/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/tex-element-image-2d-basic-expected.txt


FAIL verify the texElementImage2D works as expected assert_equals: bottom left of bottom left blue expected 255 but got 0
